feat: encrypted artifact actions
Instead of using a base64 encoded file, which is not secure, this PR uses encrypted artifacts to be able to share one or more files containing credentials with the reusable workflow.

BREAKING CHANGE: the base64 encoded credential file is not longer supported and needs to be replaced with the encrypted artifact upload / download actions.